Improve solution structure
