+++ Examples/MinimalExample/Program.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using ConvNetSharp;
+using ConvNetSharp.Layers;
+using ConvNetSharp.Training;
 
 namespace MinimalExample
 {
