+++ src/ConvNetSharp/ConvNetSharp.csproj
@@ -45,34 +45,35 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AdadeltaTrainer.cs" />
-    <Compile Include="AdagradTrainer.cs" />
-    <Compile Include="AdamTrainer.cs" />
-    <Compile Include="ConvLayer.cs" />
-    <Compile Include="DropOutLayer.cs" />
-    <Compile Include="FullyConnLayer.cs" />
-    <Compile Include="IClassificationLayer.cs" />
-    <Compile Include="IDotProductLayer.cs" />
-    <Compile Include="ILastLayer.cs" />
-    <Compile Include="InputLayer.cs" />
-    <Compile Include="LayerBase.cs" />
-    <Compile Include="MaxoutLayer.cs" />
+    <Compile Include="Layers\Activation.cs" />
+    <Compile Include="Training\AdadeltaTrainer.cs" />
+    <Compile Include="Training\AdagradTrainer.cs" />
+    <Compile Include="Training\AdamTrainer.cs" />
+    <Compile Include="Layers\ConvLayer.cs" />
+    <Compile Include="Layers\DropOutLayer.cs" />
+    <Compile Include="Layers\FullyConnLayer.cs" />
+    <Compile Include="Layers\IClassificationLayer.cs" />
+    <Compile Include="Layers\IDotProductLayer.cs" />
+    <Compile Include="Layers\ILastLayer.cs" />
+    <Compile Include="Layers\InputLayer.cs" />
+    <Compile Include="Layers\LayerBase.cs" />
+    <Compile Include="Layers\MaxoutLayer.cs" />
     <Compile Include="Net.cs" />
-    <Compile Include="NetsterovTrainer.cs" />
-    <Compile Include="PoolLayer.cs" />
+    <Compile Include="Training\NetsterovTrainer.cs" />
+    <Compile Include="Layers\PoolLayer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RandomUtilities.cs" />
-    <Compile Include="RegressionLayer.cs" />
-    <Compile Include="ReluLayer.cs" />
-    <Compile Include="SgdTrainer.cs" />
-    <Compile Include="SigmoidLayer.cs" />
-    <Compile Include="SoftmaxLayer.cs" />
-    <Compile Include="SVMLayer.cs" />
-    <Compile Include="TanhLayer.cs" />
-    <Compile Include="TrainerBase.cs" />
+    <Compile Include="Layers\RegressionLayer.cs" />
+    <Compile Include="Layers\ReluLayer.cs" />
+    <Compile Include="Training\SgdTrainer.cs" />
+    <Compile Include="Layers\SigmoidLayer.cs" />
+    <Compile Include="Layers\SoftmaxLayer.cs" />
+    <Compile Include="Layers\SVMLayer.cs" />
+    <Compile Include="Layers\TanhLayer.cs" />
+    <Compile Include="Training\TrainerBase.cs" />
     <Compile Include="Volume.cs" />
     <Compile Include="VolumeUtilities.cs" />
-    <Compile Include="WindowgradTrainer.cs" />
+    <Compile Include="Training\WindowgradTrainer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
