+++ src/ConvNetSharp/Layers/Activation.cs
@@ -1,0 +1,11 @@
+namespace ConvNetSharp.Layers
+{
+    public enum Activation
+    {
+        Undefined,
+        Relu,
+        Sigmoid,
+        Tanh,
+        Maxout
+    }
+}
