+++ src/ConvNetSharp/Net.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using ConvNetSharp.Layers;
 
 namespace ConvNetSharp
 {
